+++ test/src/i2cNoteQueryLength_test.cpp
@@ -23,7 +23,7 @@
 DEFINE_FFF_GLOBALS
 FAKE_VALUE_FUNC(const char *, noteI2CReceive, uint16_t, uint8_t *, uint16_t,
                 uint32_t *)
-FAKE_VALUE_FUNC(long unsigned int, NoteGetMs)
+FAKE_VALUE_FUNC(uint32_t, NoteGetMs)
 
 namespace
 {
@@ -55,7 +55,7 @@ SCENARIO("i2cNoteQueryLength")
 {
     NoteGetMs_fake.custom_fake = NoteGetMsIncrement;
     uint32_t available = 0;
-    size_t timeoutMs = 5000;
+    uint32_t timeoutMs = 5000;
 
     GIVEN("noteI2CReceive returns an error") {
         noteI2CReceive_fake.return_val = "some error";
