+++ test/src/i2cNoteTransaction_test.cpp
@@ -24,9 +24,9 @@ FAKE_VALUE_FUNC(void *, NoteMalloc, size_t)
 FAKE_VOID_FUNC(NoteLockI2C)
 FAKE_VOID_FUNC(NoteUnlockI2C)
 FAKE_VALUE_FUNC(const char *, i2cChunkedTransmit, uint8_t *, uint32_t, bool)
-FAKE_VALUE_FUNC(const char *, i2cNoteQueryLength, uint32_t *, size_t)
+FAKE_VALUE_FUNC(const char *, i2cNoteQueryLength, uint32_t *, uint32_t)
 FAKE_VALUE_FUNC(const char *, i2cChunkedReceive, uint8_t *, uint32_t *, bool,
-                size_t, uint32_t *)
+                uint32_t, uint32_t *)
 
 namespace
 {
@@ -41,7 +41,7 @@ SCENARIO("i2cNoteTransaction")
     char *rsp = NULL;
     const char *err = NULL;
     NoteMalloc_fake.custom_fake = malloc;
-    size_t timeoutMs = CARD_INTER_TRANSACTION_TIMEOUT_SEC;
+    uint32_t timeoutMs = CARD_INTER_TRANSACTION_TIMEOUT_SEC;
 
     GIVEN("i2cChunkedTransmit returns an error") {
         i2cChunkedTransmit_fake.return_val = "some error";
@@ -83,7 +83,7 @@ SCENARIO("i2cNoteTransaction")
     GIVEN("i2cNoteQueryLength reports bytes are available to read from the"
           " Notecard") {
         i2cNoteQueryLength_fake.custom_fake = [](uint32_t * available,
-        size_t) -> const char* {
+        uint32_t) -> const char* {
             *available = ALLOC_CHUNK;
 
             return NULL;
@@ -120,7 +120,7 @@ SCENARIO("i2cNoteTransaction")
                   "the Notecard") {
             // Write out the number of bytes reported available on the prior
             // call to i2cNoteQueryLength and report no more bytes available.
-            i2cChunkedReceive_fake.custom_fake = [](uint8_t *buf, uint32_t *size, bool, size_t,
+            i2cChunkedReceive_fake.custom_fake = [](uint8_t *buf, uint32_t *size, bool, uint32_t,
             uint32_t *available) -> const char* {
                 memset(buf, 'a', *available - 1);
                 buf[*available - 1] = '\n';
@@ -162,7 +162,7 @@ SCENARIO("i2cNoteTransaction")
             // On the first call to i2cChunkedReceive, a string of a's is
             // written to the output buffer, and the Notecard reports that there
             // are still ALLOC_CHUNK bytes to read.
-            auto firstChunk = [](uint8_t *buf, uint32_t *size, bool, size_t,
+            auto firstChunk = [](uint8_t *buf, uint32_t *size, bool, uint32_t,
             uint32_t *available) -> const char* {
                 memset(buf, 'a', *available);
                 *size = *available;
@@ -173,7 +173,7 @@ SCENARIO("i2cNoteTransaction")
             // On the second call, the remaining ALLOC_CHUNK bytes are written
             // out, which is a string of a's terminated with a newline.
             // available is set to 0, indicating there's nothing left to read.
-            auto secondChunk = [](uint8_t *buf, uint32_t *size, bool, size_t,
+            auto secondChunk = [](uint8_t *buf, uint32_t *size, bool, uint32_t,
             uint32_t *available) -> const char* {
                 memset(buf, 'a', *available - 1);
                 buf[*available - 1] = '\n';
@@ -183,7 +183,7 @@ SCENARIO("i2cNoteTransaction")
                 return NULL;
             };
             const char *(*recvFakeSequence[])(uint8_t *, uint32_t *, bool,
-                                              size_t, uint32_t *) = {
+                                              uint32_t, uint32_t *) = {
                 firstChunk,
                 secondChunk
             };
